Match the layout of ENetAddress in enet.h
The managed `ENetAddress` structure did not match the layout of the native
`ENetAddress` structure defined in enet.h.

The native structure has an additional field, so add that same field to
the corresponding managed structure.

Without this field in the managed structure, p/invoke calls into the
native code will marshal it with some arbitrary value in the native side
of the field, which can lead to unexpected problems. This happens with
Unity's IL2CPP, for example.

Fixes #13